requirements.txt
aiomysql==0.1.1
alembic==1.8.1
anyio==3.6.1
atomicwrites==1.4.1
attrs==22.1.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
fastapi==0.79.0
firebase-admin==5.3.0
google-api-core==2.8.2
google-api-python-client==2.60.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-service-directory==1.5.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
kazoo==2.8.0
Mako==1.2.1
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-connector-python==8.0.21
packaging==21.3
pluggy==1.0.0
proto-plus==1.22.1
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-multipart==0.0.5
redis==4.1.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
uvicorn==0.18.2
wrapt==1.14.1
